Added a README file.
File's contents are based upon the large comment block previously present at the start of the file `system_info_retrieval.ps1`.

File's contents about what information this utility does retrieve is not (as of this commit) properly indicative of what information is retreived by the utility; this will change in the following commits.